leon: Adjust memoryview_gc test so it runs under the leon emulator.
It looks like there is a bug in leon2-emu because running this test with a
range of 10000 leads to the emulator getting stuck on a particular
instruction.  The PC is 400397b4 in the code below and that indicates that
the write to PSR is problematic.  The code seems to be within the RTEMS
task switching routine.

400397a4:       81 8c 00 00     mov  %l0, %psr
400397a8:       01 00 00 00     nop
400397ac:       01 00 00 00     nop
400397b0:       01 00 00 00     nop
400397b4:       a9 50 00 00     rd  %wim, %l4