leon: Remove dependency on non-freely accessible code.
Due to licensing concerns, we cannot use any of the code in the "support"
folder of the EDISOFT compiler. After doing some tests, it seems that
the only pieces of that code that we depended on were the ones used to
perform output from the serial port. This set of patches:

- Provides a minimal implementation of outputing a character on
  leon2's first serial port
- Patches mphalport to delegate to this function.
- Removes the dependencies on the support/ folder from the relevant
  Makefiles

This patch also changes the default toolchain to EDISOFT RTEMS 4.8.