leon: tests: Update NOHEAP_05 to work with new float parser.
The new float parser changed how some floats are parsed and then they are
printed with a small (1e-16) difference.  So this test is updated so that
it uses floats which print without lots of 0's, and to use round() where
this is not easy to do.  This test now prints the same output with CPython,
unix uPy and the LEON port.