leon: tests: Update ROB_STACK tests to work with new pystack.
With pystack enabled these stack tests change behaviour, mostly because the
maximum stack depth is changed.  For simple recursion it's now possible to
have much deeper recursion.  For recursive functions that use a lot of
local state in the function, the recursion is less because these functions
now allocate purely on the pystack, rather than using the heap for the
function call frame.  For the big exception test (ROB_STACK_04) the test
now runs much better.