leon: Make sparc_install_ta_3_window_flush_isr always available.
The header sparcisr.h is part of the ABI and may be included by code which
does not define RTEMS_4_8_EDISOFT yet uses 4.8 Edisoft.  In such cases the
definition of sparc_install_ta_3_window_flush_isr will be inconsistent.  So
just make this function always available, and provide the correct
definition in the .c file (which must have RTEMS_4_8_EDISOFT defined if
that toolchain is used).